compilation and warnings fixes in mac

+++ cpp/file_locking/file_locking.cc
@@ -29,19 +29,19 @@ int main(int argc, char *argv[])
 
 printf("1.\n");
 
-    printf( "sizeof void*     %d\n", sizeof(void*) );
+    printf( "sizeof void*     %ld\n", sizeof(void*) );
 
 printf("2.\n");
 
-    printf( "sizeof short     %d\n", sizeof(short) );
-    printf( "sizeof int       %d\n", sizeof(int) );
-    printf( "sizeof long int  %d\n", sizeof(long int) );
-    printf( "sizeof size_t    %d\n", sizeof(size_t) );
-    printf( "sizeof off_t     %d\n", sizeof(off_t) );
+    printf( "sizeof short     %ld\n", sizeof(short) );
+    printf( "sizeof int       %ld\n", sizeof(int) );
+    printf( "sizeof long int  %ld\n", sizeof(long int) );
+    printf( "sizeof size_t    %ld\n", sizeof(size_t) );
+    printf( "sizeof off_t     %ld\n", sizeof(off_t) );
 
 printf("3.\n");
 
-    printf( "max off_t can index %.1f EB (1 ExaByte = 1x10^18 Bytes )\n", pow(2,sizeof(off_t)*8) / 1E18  );
+    printf( "max off_t can index %.1f EB (1 ExaByte = 1x10^18 Bytes )\n", std::pow(2., (int) sizeof(off_t)*8) / 1E18  );
 
 printf("4.\n");
 
@@ -52,7 +52,7 @@ printf("4.\n");
 
     assert( nbpages_2gb * pagesize == LIMIT_2GB );
 
-    printf( "The 2GB limit is actually %d pages\n", nbpages_2gb );
+    printf( "The 2GB limit is actually %ld pages\n", nbpages_2gb );
 
     size_t size_2g = ( nbpages_2gb+1 ) * pagesize;
 
@@ -82,6 +82,6 @@ printf("5.\n");
     fl.l_type   = F_UNLCK;
     fcntl(fd, F_SETLK, &fl);
 
-	return 0;
+        return 0;
 }
